requirements.txt
# ── Web framework ────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4

# ── Transcription ────────────────────────────────────────────
faster-whisper==1.1.1

# ── Document parsing (agenda upload) ─────────────────────────
python-docx==1.1.2

# ── Auth & sessions ──────────────────────────────────────────
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
itsdangerous==2.2.0

# ── Email (password reset) ───────────────────────────────────
resend==0.7.2

# ── HTTP client (used by resend) ─────────────────────────────
requests==2.32.3